Refactor into two modules
The cffi.rs module provides exclusively calls to be used on the C side. For instance in a SOP
implementation.
The interop.rs module provides utilities for the user to convert HDK specific and cffi specific C structs into safe rust structs.

C ffi leakage is not prevented but is discouraged except for POD types.