Localize constraint cache
This commit adds more documentation but more importantly, the constraint
cache is updated to change by how many points are in proximity to a
neighbour. If there are no neighbours near a query point, then the
potential will not even be computed on it. This is important for dynamic
potential fields.