Fixed bug in caching + added parallel sample iters
Cachine neighbours of query points didn't allocate space for points
without neighbours, however some code expected this to be true.
The simple implementation simply uses space for each point regarless of
whether it has neighbours or not. This is simpler than having neighbours
be a sparse vector and it helps with implementing parallelization later
(I hope).