[implicits] Fixed remaining hessian bugs + new bg
The new background is from input that is not mixed with the local field,
which means it is just sets values outside the local area to be whatever
they were from the input with a background weight of 1. This is possibly
how the None bg should have been implemented. The difference is subtle
and it is only seen when the max_step is non-zero.