[softy] Bug fixes + refactor + other improvements
- Fixed a few crashes caused by zero radius: now it's a soft error.
- Refactored material used in the lib test.
- Added a flood fill algorithm that walks through the edges copying an
attribute value through the vertices (currently unused though tested).
- Added a simple background field for smooth contact constraint that is
positive radius on the outside and negative on the inside.
- Debug functionality to print jacobian and hessian condition numbers.
- Converged on the scaling method to be gradient scaling.
- Fixed constraint violation handling at the end of the timestep: This
is caused by inaccurate neighbourhood estimate from the previous
timestep. It is updated with the neighbourhood at the end of the
timestep and the simulation is restarted (it's possible I may have to
iterate on this approach (maybe increase max step instead? currently no
issues though)
- Fix remaining tests (push and bounce tests)
- Improved the interface to allow setting mu strategy and max gradient
scaling parameter dynamically.