[implicits] Finish surface hessian implementation
This concludes the surface hessian implementation (i.e. hessian with
respect to surface vertices) including the background hessian. There is
probably a lot of opportunities for optimization here.
The following list breaks down the changes:

- Implement background hessian (wrt surface vertices)
- Implement tests for background hessians
- Refactor num_surface_hessian_product_entries to reuse code with the
surface_hessian_product_indices_iter
- Implement surface_hessian_product_indices_iter allowing us to generate
indices using an iterator instead of having to pass in mutable slices.
- Reuse the code from surface_hessian_product_indices_iter in
surface_hessian_product_indices where we do pass in mutable slices since
this is what is currently being used in simulation
- Refactor face_hessian_indices_iter to include background hessian
indices as well as distribute lower triangular *sample* hessian to the
lower triangular *vertex* hessian for background as well as main sample
hessians.
- Refactor main hessian to output lower triangular part only. This saves
on about half the computation time on computing this hessian.
- Extract print_full_hessian function to be used for debugging in
background field hessian code
- Refactor hessian tests to reduce repeated code significantly.
- Fix sample_hessian_tester to also include the upper triangular part of
the hessian when testing.
- Add query hessian test ( this is for implementing the query hessian,
not to be confused with the surface hessian ). This is work in progress.