[softy] More contact constraints + improve solver
This commit is responsible for implementing the implicit surface contact
constraint where the collision object is represented by an implicit
surface instead of the deforming object as before. Here the previously
implemented contact constraint is renamed to PointContactConstraint to
indicate that the deforming object collides against points and the new
contact constraint is named ImplicitContactConstraint to indicate that
the deforming object collides against a rigid implicit surface.

In addition this commit improves the criteria by which the solver
increases/decreases the max_step. It is still not perfect but better
than before because it doesn't assume that gradient is necessarily
scaled to max_gradient_scaling.