[implicits] Use radius_multiplier * base_radius
Instead of setting an absolute radius on the implicit surface explicit,
we instead opt to set a radius multiplier and determine a base radius
based on the input mesh. In particular we use the minimum radius that
covers each triangle in the mesh. This is more appropriate for face
centric implicit surface, but should also work for vertex centric ones
as well.