[softy] Constraints refactor for error and kernels
Refacctor constraint code to propagate errors better which allows errors
from the implicits crate to propagate down the chain. This refactor
removes crashes that were caused by a bug in implicits (which is now
fixed). Instead an error is returned an Ipopt simply ignores the values
for the constraints (in that case it was the hessian), so the derivative
test would fail but otherwise no crash.

Also refactored passing radius and threshold to constraints. This allows
us to test with various kernels instead of the Approximate one, which I
found is probably over kill when the radius is small.