[softy] Major refactor displacements -> velocities
This commit moves the optimization to use velocities instead of
displacements and force impulses instead of forces.
The rationale is that before implicit integration was implemented in a
number of different places in the code to make displacement based
optimization work. Now integration is done in one (maybe two) place(s)
in problem.rs.

To avoid performance and accuracy regressions from the previous
implementation, a variable scaling feature was added along-side this
refactor. This allows us to scale variables as if they were
displacements making all the tests consistent with the displacement
formulation. Currently the scaling parameter is simply the reciprocal of
dt. This actually makes a lot of code redundantly compute a 1.0 scaling
since in many places dt is multiplied by the scale. However, I am
leaving it as is because in general scaling can be changed to get better
convergence for more important examples in the future.