[implicits] Attempt to fix projection bias
The previous implementation of projection chose an arbitrary direction
to project when the derivative is zero. This is not ideal for cases
where the field has a constant global potential. Instead we opt to not
project when a query point is outside the local region. Although this
will not project all points out, it will prevent artifacts from forming
in the result.