[implicits] Fix projection bug
Since we didn't update the neighbourhood data when projecting vertices
along the implicit field, it was possible to project a vertex to a
completely different neighbourhood, which may produce a further
adjustment to the projection step.
This commit invalidates the nieghbourhood between projection steps. This
adds a performance penalty for reconstructing the rtree between
projections, however this overhead doesn't seem to be significant.