[softy] Reorganizing objects into solids and shells
This commit is very much a WIP save.

Overview:
TetMeshes and Kinematic collision objects are being generalized into
solids and shells. Each of these get their own material properties in
preparation for simulating multiple bodies as well as shells (cloth).
The next step is to organize how a pair of objects is assigned
frictional contact parameters such that a one(or two)-way constraint is
added in the appropriate place in the simulator.

The Softy API for determining frictional contacts is as follows:
 1. By default all objects are decoupled: no contact (no default
    frictional contact params)
 2. Solids are allowed to have variable material properties including
    friction. This means that for a given homogeneous material X, a
    solid may define a variable coefficient mu that defines its
    coefficient of friction with X.
    Typically X is a cloth (shell), but it could be another solid also.
 3. If no variable friction attribute is found, use globally specified
    parameters (via the add_frictional_contact function).