[utils] impl of Index<usize> + ToOwned + tests
Added implementations of Index<usize> for Chunked and UniChunked.
Added a custom ToOwned trait to supplement the one used by the standard
library (since it expects borrows).
Added doc tests for the new implementations as well as a
zero_length_chunk test for Chunked where some of the chunks have zero
length.