[utils] Add Subset::from_unique_order_indices
This constructor only expects that the given set of indices have the
correct form (sorted and unique), and panics if it doesn't instead of
trying to sort and dedup it manually like in Subset::from_indices.

This makes for more efficient Subset construction in cases where the
indices are known to be valid (which is often).