[utils] Refactored Get + SubsetIterMut + more tests
Refactored Get traits to look more like the std lib get calls, which return
options instead of panicking on out of bounds. To mimic the behaviour of the
behaviour of `index` in std lib, `at` was added`. Possibly in a later revision,
at will be used as the dereferenced version of index which would mimic the []
operator in the std lib.

Additionally the SubsetIterMut was fixed to allow mutable iteration of subsets.

SplitFirst and SplitPrefix traits were introduced to make the above happen.

Additional tests were added to check composition of subsets and other chunked
types.

Also added convenience offset and offset_value getters for `Chunked` types.
These allow users to peek in to the structure of the chunking.