[utils] Rewrote Get traits to fix borrowing
The issue was that we tried to use the same getter trait to index both
owned and borrowed (viewed) collections with a trait like:

```rust
trait Get<'i, 'o> {
    type Output;
    fn get(&'i self, idx: usize) -> Option<Self::Output>;
}
```

Where 'o is intended for the output lifetime and 'i outlives 'o
(in other words `'i: 'o`).

This works well for owned collections where 'o is guaranteed not to
outlive the data in the owned collection. However, for viwed collection,
the true borrow lifetime lives inside Self, and 'i really ties the
output 'o to the borrow of the view (which is itself a borrow). But the
output is a reborrow of the view and can outlive the view itself.
Encoding this logic in lifetimes proved to be difficult, so the code was
rewritten and simplified with a few more unsafe calls.
The Get traits were rewritten with a single lifetime for the output and
an arbitrary lifetime for self, allowing views to be reborrowed. This
also means that they cannot be used with owned collections.

Maybe at a later time this can be improved, but for now it is sufficient
to make the rest of the code work.