[utils] Implement subset of subsets + tensor ergonomics
This commit enables us to have subsets of subsets.
In addition the as_tensor function is moved to as_mut associated function of
Tensor and a corresponding non-mutable variant is made called as_ref. These
should make it easier to convert types to their tensor representations even if
we are not allowed to move these types. This is useful because tensor is meant
to act as an ephemeral type to allow computations on structured data.