[utils] Rework UniChunked Elem
Previously UniChunked elem was define to be an array.
This is inconsistent with composite types and the types used in iteration.
Instead of array, a new trait UniChunkable was introduced to put the burden of
defining Elem of UniChunked on the underlying collection.
In addition, PushChunk was introduced to let collections define how to push a
chunk onto them.
Finally CloneIntoOther was separated into a trait from Subset to allow composite
definitions of this behaviour with other types.
Although due to lifetime issues, clone_into_other is defined on the subset type
directly.

This work allowed us to implement conversion from BlockDiagonalMatrix2 to
DSBlockMatrix2 more easily.