[implicits] Major refactor of implicits package
This commit includes many changes to make the implicits package more
declarative instead of object oriented.

The goal is to reduce the need for interior mutability as to simplify
the interfaces for getting iterators for jacobians and hessians more
easily.

This refactor is a step towards decoupling various types of implicit
surfaces at the cost of some api duplication, which is meant to be
abstracted into traits at a later time.