[implicits] Moving to using QueryTopo struct
This commit implements a QueryTopo struct that encapsulates the
information about the topology between a set of query points an the
implicit surface defined by a set of samples. It is now recommended to
use this struct instead of running compute_neighbourhoods on the
implicit surface directly.

The reasoning behind why we want to do this is it makes it explicit and
simple to reason about the state of the implicit surface. It should
further be possible to compute potentials and derivatives of the
surfaces without requiring precomputations of query topology --- for
instance we may use the spatial tree directly to iterate over neighbours
when computing derivatives instead of relying on an auxiliary table of
neighbours being precomputed beforehand.