[implicits] Removed chunking + benchmark
Benchmarked how chunking affects performance of work-stealing algorithm
in rayon. Turns out chunking the main potential computation doesn't add
any performance benefits, so that code was removed.
It is unclear if the chunking was done incorrectly, or simply doesn't
mesh well with the way rayon parallelizes. Either way the current state
is the simplest, so we will keep it until further investigation is
needed.

A benchmark was added to test the performance of a simple torus against
a grid plane.