[implicits] Reduce compile times + prune dead code
Feature gated Serde to prevent unnecessary codegen.
Remove unused code for computing on the MLS struct directly. This may be
reintroduced but in a different way that allows computing potentials
without having to build a QueryTopo struct.
Switched to Box<dyn Iterator> instead of impl Iterator for a few
neighbourhood accessors in hopes of reducing compile times.