[utils] Grand nested tensor refactor
This refactor changes tensor types to be interleaved when converting to
tensors from composed collection types.

This was done to ease some of the difficulties in specifying trait
bounds on ops implementations for tensors.

This seems to have worked, although a lot of functionality has been
pruned.

AFAICT this mostly affects eager (non-lazy) ops on tensors.
A significant performance regression is noticed for lazy adds and matrix
multiplies in the benchmarks which may be due to the pruning (not
necessarily related to the refactor itself).