[softy] Fix ref pos attrib + remove material normalization
Reference position was not properly output in houdini because the name
"ref" clashed between vertex and face-vertex attributes.
This is not an issue outside of houdini.

Material normalization was removed since it no longer makes sense to
normalize materials without taking into account all interacting
materials together. In addition the problem is already being scaled in
fem/problem.rs, so this should now be unnecessary anyways.