[softy] Substantial refactor of constraints
This refactor should make integrating rigid bodies more seamless with
the current constraint system.

Merged LinearizedPointContact into PointContact
Made FrictionalContacts non trait objects but all PointContacts since
this is the only thing we have so far. In the future this may be changed
back to trait objects but
 1. For the time being the extra complexity is unnecessary.
 2. The constraint interface needs to be completely rewritten anyways
since there is a lot of old cruft still lingering there.

Bulding frictional contact pairs is refactored. This was a big ugly
function that does very little and it is now made substantially
simpler.