[softy-hdk] Add use fixed parameter for contacts
This is intended to enable/disable the use of fixed element faces to
generate contact surfaces.

In a standard pipeline, the user may choose to remove animated/fixed
elements from the mesh. This would typically cause the solver to
generate a collision surface on the inside of the mesh. This flags is
used to ignore these types of "interior" surfaces when generating the
collision surface.