[softy] Added contact_offset and use_fixed contact params
Contact offset sets the iso-value of the implicit surface at which to limit interpenetration.
In other words this is the distance from the zero implicit surface at which any colliding points will stop.
In effect this allows thin shells to look better when colliding against solids.

Use fixed is currently a dummy parameter, which is intended to be used to prune collision geometry on partially fixed meshes.