[softy] Format + refactor bending energy
Added a dedicated Bending energy struct for computing per interior edge
bending energies. This commit includes an alternative Tan discrete
energy to see what a tan energy would produce. This was motivated by
wierd kinks found in the cloth.

It turns out that the kinks are characteristic of the bending stiffness
being too high (in comparison to stretch stiffness).  It is not clear
how to fix this, but a tan energy isn't it. This commit leaves that
energy for future testing and reference. Although it is not being used.