[implicits-hdk] Improve interface
Refactored the C facing api and added base radius capability.
This allows us to construct an implicit field with an absolute radius
rather than one relative to the size of the mesh, which is useful when
comparing implicit fields generated by the simulation, which keeps the
base radius constant throughout the sim.