[softy] Iterate on NL friction
This commit introduces a pseudo working state for penalty based contact
with friction. This is still a prototype and the linear solve is still
lacking, but it may just be a configuration issue.
More debugging is needed to determine performance bottlenecks and
other configuration issues.
Not all NL tests are passing.