[softy] WIP fixing Newton solve
Fixed an incorrect armijo condition check and other tweaks.
Left a bunch of debug code to be removed later.

Friction needs to be reworked to prevent contact pruning. This doesn't
work with nl friction since contacts can come in and out during one time
step since verything is implicit, so the jacobian can change sparsity.
This is currently causing the solve to break and the derivatives to be
incorrectly computed.

Next steps is to rework the friction_jacobian function in
penalty_point_contact.

This commit marks a spot where the code seems to be doing the right
thing for the iterative solve now, but not yet for the direct solve,
which we still need for fast small scale sims.

Additionally this commit includes an rough implementation for 0.5|J^{-1}r|^2
merit function, which should help with larger time steps or stiffer systems,
we will see after we fix the friction jacobian.