[softy] Fix final friction jacobian for implicits
This commit finalizes the friction jacobian implementation including the
jacobian with respect to implicit surface, which means that we can now
use the direct method with deformable objects with implicit collision
surfaces.

This commit also includes a few straightforward optimizations of this
derivative and a small clippy cleanup.

The friction derivative is still very far from being efficient.