[softy] Implemented SDIRK2 and TRBDF2 integrators
This commit also changes how the damping parameter is used. It is now
expected to be unitless, whereas earlier it had units of area per
second as kinematic viscosity coefficient does. This should help make
the coefficient more stable to changes to the damping model.

This commit also refactors the residual computation using automatic
differentiation.