[softy] Added deterministic feature moved ref_tri
Instead of conditionally compiling determinism for interior edge
construction by checking debug build, we use a "deterministic" feature
flag to enable determinism if required. The principle is that if one
dependent needs or relies on determinism then all instances of this
dependency will have determinism. Non-deterministic solves may be
desirable at some point in the future if this library is ever exposed as
a service to avoid DDos type attacks. This is not currently the focus.

Also moved ref_tri to live next to ref_tet in src/fem/mod.rs