[softy] Small upgrades
Start on code using projected hessians. This is a gateway to
implementing energy minimization with lagged friction.

Added code to conditionally enable the energy potential used as a merit
function when both lagged friction and projected hessians are enabled.

Using only normal velocity for max step updates. This mechanism is
broken anyaways, but this prevents it from misfiring more frequently.

Output 64 bit vertex positions as an attribute, this allows 64 bit
positions to be used in Houdini, since they use 32 bits for vertex
positions in meshes.

Set background potential radius to be positive for contact constraint.
This fixes the discontinuity in the potential when max_step > 0.

Return an error when Jacobian index updates fail instead of panicking.

Added support for doing contact assisted linesearch without friction
assisted. This is done for testing and comparison.