[implicits] Fix the contact jacobian singularity
This commit resolves the singularity in the contact jacobian that occurs
when a vector on the surface is mapped to another vector on a surface
facing in the opposite direction (dot product of the normals between two
surfaces is close to -1). This can cause a discontinuity in the contact
jacobian and thus breaking the smoothness assumption of contacting
bodies.

This fix is not in the SIG 2021 paper, but a new contribution.